Fully implemented xml marshalling/unmarshalling of entities.
- jaxb.index file used to specify components that can be stored in xml;
must be placed in the same package as the components it specifies and
the package name itself must be passed to the EntityFileParser.
- Simplified entities a fair bit to have only the bare minimum required;
further customization can and should be done in the declaring classes
(e.g. giving player the ability to be user controlled and gui).
- EEntity class now has objects that reference file names of their
Entity, which should be located in the same package name as EEntity and
should also be in the resource source file.
- EntityBlueprintFileMaker can be used to easily create and save entity
blueprint files; note that they are saved in the res directory and have
to be manually moved into the classpath, which, as of now, is the only
place they are loaded from.