Changed how states are transitioned from/to.
- Made Client use GridPane as root, with the root of each State
occupying a single cell in that so that transitioning from one state to
the next would be much easier. This didn't work too well as the GridPane
needs to be scaled and that messes up how the States look.
- Added BlendMode SRC_OVER to UserComponentRender which should cause
player names to always be visible, even when on a background of the same
color.